Move temp cloud storage to storage
