+++ lib/engine.js
@@ -22,7 +22,7 @@ const Asset = require('./sdk/asset');
 const Rendition = require('./sdk/rendition');
 const {PLAN_STATE} = require('./plan');
 const { getSource, putRendition } = require('./sdk/storage');
-const {TemporaryCloudStorage} = require('./temporary-cloud-storage');
+const {TemporaryCloudStorage} = require('./storage/temporary-cloud-storage');
 const {GenericError, AssetComputeEvents, AssetComputeMetrics, OpenwhiskActionName, Reason, ClientError, AssetComputeLogUtils} = require('@adobe/asset-compute-commons');
 
 // get utils
