+++ lib/sdk/storage/temporary-cloud-storage.js
